크기 6의 덱 시스템 생성, 왼클릭하면 현재 선택된 덱의 유닛을 소환함

+++ GPOS Winter Project 2019/Assets/Scripts/Projectile/Arrow.cs
@@ -22,7 +22,7 @@ public class Arrow : MonoBehaviour
         {
             if (!collision.gameObject.tag.Equals(team.ToString()))
             {
-                collision.gameObject.GetComponent<Unit>().Damage((uint)damage);
+                collision.gameObject.GetComponent<Unit>().Damage(damage);
                 Destroy(gameObject);
             }
         }
