+++ GPOS Winter Project 2019/Assets/Scripts/Projectile/Explosive.cs
@@ -32,7 +32,7 @@ public class Explosive : MonoBehaviour
                     else
                     {
                         Debug.Log(i.ToString() + " : " + Targets[i].gameObject.ToString());
-                        if(Targets[i].gameObject.GetComponent<Unit>().TeamTag != team) Targets[i].gameObject.GetComponent<Unit>().Damage((uint)damage);
+                        if(Targets[i].gameObject.GetComponent<Unit>().TeamTag != team) Targets[i].gameObject.GetComponent<Unit>().Damage(damage);
                     }        
                 }
                 Destroy(gameObject);
